Create construct module
House the world and world construction components in their own module
