+++ stress_fiber/construct/__init__.py
@@ -1,0 +1,2 @@
+from .world import World  # noqa: F401
+from .test import create_test_world  # noqa: F401
